Move methods to MessageFactory.
