+++ registration/src/test/java/org/spine3/samples/lobby/registration/conference/ConferenceProjectionRepositoryShould.java
@@ -25,12 +25,12 @@ import org.junit.Before;
 import org.junit.Test;
 import org.spine3.samples.lobby.common.ConferenceId;
 import org.spine3.samples.lobby.conference.contracts.Conference;
-import org.spine3.samples.lobby.registration.testdata.TestDataFactory;
 import org.spine3.server.storage.EntityStorage;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
 
 import static org.junit.Assert.assertEquals;
 import static org.spine3.samples.lobby.registration.testdata.TestDataFactory.newBoundedContext;
+import static org.spine3.samples.lobby.registration.util.MessageFactory.newConferenceId;
 
 /**
  * @author Alexander Litus
@@ -39,7 +39,7 @@ import static org.spine3.samples.lobby.registration.testdata.TestDataFactory.new
 public class ConferenceProjectionRepositoryShould {
 
     private final ConferenceProjectionRepository repository = new ConferenceProjectionRepository(newBoundedContext());
-    private final ConferenceId id = TestDataFactory.newConferenceId();
+    private final ConferenceId id = newConferenceId();
 
     @Before
     public void setUpTest() {
