+++ registration/src/test/java/org/spine3/samples/lobby/registration/testdata/TestDataFactory.java
@@ -23,17 +23,10 @@ package org.spine3.samples.lobby.registration.testdata;
 import org.spine3.base.EmailAddress;
 import org.spine3.base.PersonName;
 import org.spine3.eventbus.EventBus;
-import org.spine3.money.Currency;
-import org.spine3.money.Money;
-import org.spine3.samples.lobby.common.ConferenceId;
-import org.spine3.samples.lobby.common.OrderId;
 import org.spine3.samples.lobby.common.PersonalInfo;
-import org.spine3.samples.lobby.common.SeatTypeId;
-import org.spine3.samples.lobby.registration.contracts.SeatQuantity;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.CommandDispatcher;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
-import org.spine3.util.Identifiers;
 
 /**
  * The utility class which is used for creating objects needed in tests.
@@ -60,39 +53,6 @@ public class TestDataFactory {
     }
 
     /**
-     * Creates a new {@link OrderId} with a random UUID value.
-     */
-    public static OrderId newOrderId() {
-        final String id = Identifiers.newUuid();
-        return OrderId.newBuilder().setUuid(id).build();
-    }
-
-    /**
-     * Creates a new {@link ConferenceId} with a random UUID value.
-     */
-    public static ConferenceId newConferenceId() {
-        final String id = Identifiers.newUuid();
-        return ConferenceId.newBuilder().setUuid(id).build();
-    }
-
-    /**
-     * Creates a new {@link SeatTypeId} with a random UUID value.
-     */
-    public static SeatTypeId newSeatTypeId() {
-        return SeatTypeId.newBuilder().setUuid(Identifiers.newUuid()).build();
-    }
-
-    /**
-     * Creates a new {@link Money} instance with the given {@code amount} and {@code currency}.
-     */
-    public static Money newMoney(int amount, Currency currency) {
-        final Money.Builder result = Money.newBuilder()
-                .setAmount(amount)
-                .setCurrency(currency);
-        return result.build();
-    }
-
-    /**
      * Creates a new {@link PersonalInfo} instance with the given {@code givenName}, {@code familyName} and {@code email}.
      */
     public static PersonalInfo newPersonalInfo(String givenName, String familyName, String email) {
@@ -102,14 +62,4 @@ public class TestDataFactory {
         return result.build();
     }
 
-    /**
-     * Creates a new {@link SeatQuantity} instance with the given {@code quantity} and random UUID.
-     */
-    public static SeatQuantity newSeatQuantity(int quantity) {
-        final String id = Identifiers.newUuid();
-        final SeatQuantity.Builder result = SeatQuantity.newBuilder()
-                .setSeatTypeId(SeatTypeId.newBuilder().setUuid(id))
-                .setQuantity(quantity);
-        return result.build();
-    }
 }
